修复：数据库模型中 Rating 不是浮点数的问题​

+++ server/database/models/Item.js
@@ -96,7 +96,7 @@ Rating.init({
     defaultValue: 0
   },
   rating: {
-    type: Sequelize.INTEGER,
+    type: Sequelize.FLOAT,
     allowNull: false,
     defaultValue: 0
   },
